requirements3_12.txt
pillow=11.1.0
pip=23.2.1
